+++ Firmware/src/Firmware.Core/Utilities/SerializableObject/Converters/ArrayConverter/ArrayConverter.h
@@ -16,12 +16,12 @@ namespace junjinjen_matrix
 					class ArrayConverter
 					{
 					public:
-						bool TrySetArray(std::basic_ostream<uint8_t>& stream, const T* elements, size_t elementsCount) const
+						bool TrySetArray(std::basic_ostream<uint8_t>& stream, const T* elements, int32_t elementsCount) const
 						{
 							ValueConverter<T> converter;
-							stream.write((uint8_t*)&elementsCount, sizeof(size_t));
+							stream.write((uint8_t*)&elementsCount, sizeof(int32_t));
 
-							for (size_t i = 0; i < elementsCount; i++)
+							for (int32_t i = 0; i < elementsCount; i++)
 							{
 								if (!converter.TrySetValue(stream, elements[i]))
 								{
@@ -34,11 +34,11 @@ namespace junjinjen_matrix
 
 						bool TrySetArray(std::basic_ostream<uint8_t>& stream, const std::vector<T>& elements) const
 						{
-							size_t size = elements.size();
+							int32_t size = elements.size();
 							ValueConverter<T> converter;
-							stream.write((uint8_t*)&size, sizeof(size_t));
+							stream.write((uint8_t*)&size, sizeof(int32_t));
 
-							for (size_t i = 0; i < size; i++)
+							for (int32_t i = 0; i < size; i++)
 							{
 								if (!converter.TrySetValue(stream, elements[i]))
 								{
@@ -52,14 +52,14 @@ namespace junjinjen_matrix
 						bool TryGetArray(std::basic_istream<uint8_t>& stream, std::vector<T>& elements) const
 						{
 							ValueConverter<T> converter;
-							size_t size;
-							if (!stream.read((uint8_t*)&size, sizeof(size_t)))
+							int32_t size;
+							if (!stream.read((uint8_t*)&size, sizeof(int32_t)))
 							{
 								return false;
 							}
 
 							elements.resize(size);
-							for (size_t i = 0; i < size; i++)
+							for (int32_t i = 0; i < size; i++)
 							{
 								if (!converter.TryGetValue(stream, elements[i]))
 								{
